Detection of the lack of libraries (mmap and six) in Jython.
In order to use jpylyzer in Java using Jython, the use of the mmap
library should be avoid. Indeed the mmap is not supported by Jython
(see https://bugs.jython.org/issue1854873).

To simplify the detection for the lack of libraries in Jython, the ImportError
exception is catch and a flag is set.
When the mmap library is not available, files are directly read in memory
(method used before the 1.14 version of jylyzer).

An example java program is provided to test.

Takes also care of the six library which doesn't work with Jython
Correct some E302 pep8 warnings