Add plt segment to properties output
I didn't add any validation, as for large files this could take up a significant percentage of the output file.
This will significantly increase the output size for datasets that use plt's. The output of a ~6GB file with plt's went from ~90KB to ~90MB.

The logic here could also be used to parse the plm segment. However I don't have any sample datastes that use that, so I didn't implement this.